+++ NEWS.md
@@ -2,6 +2,7 @@
 # seplyr 0.1.6 2017-08-29
 
  * work on lists.
+ * re-export more dplyr functions.
 
 # seplyr 0.1.5 2017-08-28